[Android] Adjustment to allow selecting a pane before showing HubLayout
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/4d6f6030a260d8101b5826a63dbc5261684bc512

[Hub] Allow selecting a pane before showing HubLayout.

In this CL, we add class HubShowPaneHelper. Chrome can select the next
pane id before opening the HubLayout via the HubShowPaneHelper
instance. The HubLayout will read the next pane id and set focus to the
selected pane. The HubShowPaneHelper is created by the HubProvider and
plumb to HubLayout from HubManager.

Bug: 41489743