Convert SplitView unit tests to browser tests
Upstream code chains calls to GetBrowserWindowInterface() in several places now,
preventing the use of TestTabStripModelDelegate as-is in our SplitView unit
tests. Due to ongoing upstream changes in BrowserWindowInterface, that class
also is not safely mockable as currently written. Instead, we've converted the
SplitView unit tests to browser tests so there is no longer a need to mock those
interfaces, which were only used very sparingly upstream to begin with.