[Android] Adopt ServiceLoaderUtil for LanguageProfileDelegate
Upstream's logic in `AppLocaleUtils` for `getApplicationLocales` and `setApplicationLocales` is the same as it was in our patch, so we don't need it anymore.

Chromium change:
https://source.chromium.org/chromium/chromium/src/+/826aadf32e2ebec6a4086d97937f345ea5f3f912

Android: Adopt ServiceLoaderUtil for LanguageProfileDelegate

And upstreams & inlines the logic from LocaleManagerDelegate, which
existed only to work around Android T SDK not yet being available.

Also adds some @JniType annotations to LanguageBridge to simplify
changes there.

Bug: 40901855