[wallet] Support for 64-bit integers over JSON
This PR adds a couple of changes to chromium's JSON reader/writer to
allow the use of 64-bit integers with JSON. This is currently not
supported in chromium as it violates the JS specification for integer
numbers.

With these changes, there are special flags that can be provided to
`JSONReader`, and `JSONWirter`, to allow us to treat any value that can
only be represented as `int64_t`/`uint64_t` as a binary value. This
binary value is stored in string format, which preserve the information
necessary to safely convert it to and from bytes. For values that can
fit an int, things stay the same to preserve consistency with how
`double` promotion is handled for `base::Value`

Additionally, this change adds machinery code support into the json
schema compiler, so this can be hidden away, reducing friction dealing
with with these values.

```webidl
dictionary Example {
  [override=int64_t] ArrayBuffer singedValue;
  [override=uint64_t] ArrayBuffer unsignedValue;
};
```

With the schema above, the compiler will produce a struct with fields of
types `int64_t`, and `uint64_t` respectively. It will also produce the
necessary codegen to hide away the handling of the buffer.