[Android] Added playback mic service to the exception list
Otherwise there is an error:
Service org.chromium.chrome.browser.media.ui.BraveMediaNotificationControllerServices$PlaybackListenerMicService should be declared in the base manifest, but is in "chrome" module. For details, see b/169196314.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/18745fae2496b52add5a4b678f8954c4a5b75718

Fix bug where we did not validate bundle for isolated splits

Added an allowlist to get this check back on, and then we can decide
on priority of burndown.

Bug: 381244132