[bugfix] Add `chromium_src` check heuristics to detect partial matches (#28609)
It was recently uncovered that an override `ShouldShowPromo` silently
slipped away, as the presubmit check was partially matching for
`ShouldShowPromoBasedOnImpressionOrDismissalCount`. This has been hiding
a live bug that users have reported where they were being shown the sign
in promo.

This change does a few things. First, there's an improvement to how the
presubmit works, by making sure overrides do occur in the target files
by matching them as a token.

Additionally, to determine if it is an error or a warning the absecence
of a token in the target file, the presubmit relies on a check for the
token occcurring in the override file, and when it does, it concludes
just warns, as there is a chance the macro is being used in the
override. This check is being tweaked a bit in this change, to be
constrained only to `SHOUTY_CASE` tokens, as common word tokens are
most certainly not common macros and should occur in the target file, so
their presence in the override file is immaterial to the rules.

Lastly, this change corrects a few places where overrides had slipped
away, and it changes a few overrides to their respective files, rather
than relying on replacement of transient inclusions.

In particular, the `ShouldShowPromo` override is being fixed, and these
functions are being blocked again.

Resolves https://github.com/brave/brave-browser/issues/45354

Chromium changes:
https://chromium.googlesource.com/chromium/src/+/5692313c36387f63418877d922669cc3134a18cc
https://chromium.googlesource.com/chromium/src/+/d59dbad55ace20c22aefe84aed328f83065827ad

commit 5692313c36387f63418877d922669cc3134a18cc
Author: Amelie Schneider <amelies@google.com>
Date:   Mon May 13 09:22:14 2024 +0000

    [UNO] Don't show Sign-in promo after 5 impressions/2 dismissals

    Only show the autofill sign in promo bubble 5 times per account per
    profile with 1 week delay between impressions. Do not show it again if
    it was dismissed twice.
    Also moves SyncPromoUI::ShouldShowSyncPromo to
    chrome/browser/signin/signin_promo_util.h for consistency.

    Bug: b:319411476

    Change-Id: I8de968cb5d08dcc0561f46471eab5b68af47ec02
    Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5446906
    Commit-Queue: Amelie Schneider <amelies@google.com>
    Reviewed-by: Alex Ilin <alexilin@chromium.org>
    Reviewed-by: Ryan Sultanem <rsult@google.com>
    Cr-Commit-Position: refs/heads/main@{#1299945}

commit d59dbad55ace20c22aefe84aed328f83065827ad
Author: Caroline Rising <corising@chromium.org>
Date:   Mon Apr 8 17:23:43 2024 +0000

    Clean up chrome refresh 2023 from RecentTabsSubMenuModelTests

    Change-Id: I766820c477e06f5c156cb2b4e85f3efdbb1f9ce9
    Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5435914
    Commit-Queue: Caroline Rising <corising@chromium.org>
    Reviewed-by: Darryl James <dljames@chromium.org>
    Cr-Commit-Position: refs/heads/main@{#1283927}